tests: split test_backend_df.py into per-module files (review feedback)
One backend-test file per df module -- test_backend_surfacesigma.py +
test_backend_jeans.py -- so df tests don't accumulate in a single file as
Track F grows (each later family gets its own: sphericaldf, diskdf, qdf,
streams). Tests unchanged.
